==> all_scripts/1.1 data_compilation_fd.Rmd <==
##########
Integrating beta diversity (FD) with predictor variables
##########

# Compile FD beta
```{r}
pacman::p_load(tidyverse, readxl, glue, fossil)

data_dir <- "raw data"
beta_dir <- "D:/Desktop/Beta/R/output_sor_functional_parts/"
dir.create("table", showWarnings = FALSE)

get_off <- function(x) {
  smp <- sample(nrow(x))
  x <- x[smp, smp]
  delta <- row(x) - col(x)
  x[delta < 1 | delta > 1] <- NA
  x[upper.tri(x)] <- NA
  as.data.frame.table(x, responseName = "dist") %>%
    drop_na() %>%
    filter(Var1 != Var2)
}

file_paths <- list.files(data_dir, pattern = "^[^~].*\\.xlsx$", full.names = TRUE)
dataset_names <- tools::file_path_sans_ext(basename(file_paths))
names(file_paths) <- dataset_names

predictors_input <- list()
for (i in seq_along(file_paths)) {
  df <- read_excel(file_paths[i], sheet = "coordinates")
  names(df)[1:3] <- c("site", "X", "Y")
  df <- df %>%
    mutate(site = as.character(site), X = as.numeric(X), Y = as.numeric(Y)) %>%
    drop_na()
  predictors_input[[dataset_names[i]]] <- df
}

# TP 
tp_all <- list()
for (i in seq_along(file_paths)) {
  df <- read_excel(file_paths[i], sheet = "environment")
  names(df)[1] <- "site"
  df <- df %>%
    select(site, TP) %>%
    mutate(site = as.character(site), TP = as.numeric(TP)) %>%
    drop_na()
  tp_all[[dataset_names[i]]] <- df
}

beta_paths <- list.files(beta_dir, pattern = "_functional_beta\\.rds$", full.names = TRUE)
beta_names <- tools::file_path_sans_ext(basename(beta_paths)) %>% str_remove("_functional_beta")
names(beta_paths) <- beta_names

beta_matrices <- purrr::map(beta_paths, ~ {
  x <- readRDS(.x)
  list(
    total = as.matrix(x$funct.beta.sor),
    turnover = as.matrix(x$funct.beta.sim),
    nestedness = as.matrix(x$funct.beta.sne)
  )
})

pairwise_offdiag_randomized <- purrr::map(predictors_input, function(df) {
  coords <- df[, c("X", "Y")]
  rownames(coords) <- df$site
  mat <- as.matrix(stats::dist(coords))
  purrr::map_dfr(1:100, function(i) get_off(mat), .id = "rep")
})

input_skeleton <- bind_rows(pairwise_offdiag_randomized, .id = "dataset") %>%
  select(dataset, rep, Var1, Var2)

beta_long <- map2_dfr(beta_matrices, names(beta_matrices), function(mats, name) {
  total <- as.data.frame.table(mats$total, responseName = "beta_total")
  repl <- as.data.frame.table(mats$turnover, responseName = "beta_turnover")
  rich <- as.data.frame.table(mats$nestedness, responseName = "beta_nestedness")
  reduce(list(total, repl, rich), full_join, by = c("Var1", "Var2")) %>%
    mutate(dataset = name)
})

tp_long <- map2_dfr(tp_all, names(tp_all), function(df, name) {
  df1 <- df %>% rename(Var1 = site, TP1 = TP)
  df2 <- df %>% rename(Var2 = site, TP2 = TP)
  crossing(df1, df2) %>% mutate(dataset = name)
})

coords_long <- map2_dfr(predictors_input, names(predictors_input), function(df, name) {
  df1 <- df %>% rename(Var1 = site, x1 = X, y1 = Y)
  df2 <- df %>% rename(Var2 = site, x2 = X, y2 = Y)
  crossing(df1, df2) %>% mutate(dataset = name)
})

model_data_FD <- input_skeleton %>%
  left_join(beta_long, by = c("dataset", "Var1", "Var2")) %>%
  left_join(tp_long, by = c("dataset", "Var1", "Var2")) %>%
  left_join(coords_long, by = c("dataset", "Var1", "Var2")) %>%
  mutate(
    TP_diff = abs(TP1 - TP2),
    Mean_TP = (TP1 + TP2) / 2,
    spatial.distance = purrr::pmap_dbl(list(y1, x1, y2, x2), function(lat1, lon1, lat2, lon2) {
      fossil::earth.dist(rbind(c(lat1, lon1), c(lat2, lon2)))
    }),
    rep = factor(rep),
    dataset = factor(dataset)
  ) %>%
  drop_na(beta_total, beta_turnover, beta_nestedness, TP1, TP2)

# Organism/Ecosystem variables
dataset_var <- read.csv("input_file/dataset_var.csv") %>%
  rename(dataset = Dataset)

input_info <- dataset_var %>%
  select(dataset, Organism:Ave_Forest)

model_data_FD <- model_data_FD %>%
  left_join(input_info, by = "dataset")

#write_csv(model_data_FD, "output_file/model_data_sorensen_functional_parts.csv")
```

# Compile FD beta and Climate
```{r}
library(raster)
model_data <- read_csv("output_file/model_data_sorensen_functional_parts.csv")

AT_raster <- raster('input_file/climate/wc2.1_30s_bio_1.tif') 
TS_raster <- raster('input_file/climate/wc2.1_30s_bio_4.tif')   # no use
AP_raster <- raster('input_file/climate/wc2.1_30s_bio_12.tif')  
PS_raster <- raster('input_file/climate/wc2.1_30s_bio_15.tif')  # no use
Elevation_raster <- raster('input_file/climate/wc2.1_30s_elev.tif')   # no use

coords1 <- model_data[, c("x1", "y1")]
coords2 <- model_data[, c("x2", "y2")]

AT1 <- raster::extract(AT_raster, coords1)
AT2 <- raster::extract(AT_raster, coords2)
TS1 <- raster::extract(TS_raster, coords1)
TS2 <- raster::extract(TS_raster, coords2)
AP1 <- raster::extract(AP_raster, coords1)
AP2 <- raster::extract(AP_raster, coords2)
PS1 <- raster::extract(PS_raster, coords1)
PS2 <- raster::extract(PS_raster, coords2)
EL1 <- raster::extract(Elevation_raster, coords1)
EL2 <- raster::extract(Elevation_raster, coords2)

model_data$AT <- abs(AT1 - AT2)
model_data$TS <- abs(TS1 - TS2)
model_data$AP <- abs(AP1 - AP2)
model_data$PS <- abs(PS1 - PS2)
model_data$Elevation <- abs(EL1 - EL2)

model_data <- model_data[complete.cases(model_data), ]
# write_csv(model_data, file = "output_file/model_data_sorensen_functional_parts.csv")
```

# Compile FD beta and Landuse
```{r}
pacman::p_load(tidyverse, readxl, glue, fossil, sf, raster, exactextractr, pbapply)

model_data <- read_csv("output_file/model_data_sorensen_functional_parts.csv")
colnames(model_data)[colnames(model_data) == "beta_turnover"] <- "turnover"
colnames(model_data)[colnames(model_data) == "beta_nestedness"] <- "nestedness"

set.seed(1234)
data_selected <- model_data %>%
  filter(TP1 > 0, TP2 > 0) %>%
  group_by(rep) %>%
  slice_sample(prop = 0.005) %>%
  ungroup() %>%
  dplyr::select(dataset, rep, x1, y1, x2, y2,   #info 
         Organism, Group, Ecosystem, Site_num,Area,Ave_TP,Ave_AL,Ave_AT,Ave_AP,Ave_Elevation, # Dataset-level
         Gamma_Diversity,Max_Distance,Ave_HFP,Ave_Builtup,Ave_Crops,Ave_Forest,    # Dataset-level
         TP_diff, spatial.distance, HFP, AT, TS, AP, PS,  Mean_TP,             # Site-level  
         total, turnover, nestedness)


samples_sf1 <- st_as_sf(data_selected, coords = c("x1", "y1"), crs = 4326)
samples_sf2 <- st_as_sf(data_selected, coords = c("x2", "y2"), crs = 4326)

compute_land_cover_diff <- function(raster_path, sf1, sf2, buffer_dist) {
  lu_raster <- raster(raster_path)
  
  process_point <- function(point_geom) {
    lon <- st_coordinates(point_geom)[1]
    lat <- st_coordinates(point_geom)[2]
    utm_zone <- floor((lon + 180)/6) + 1
    utm_crs <- st_crs(glue("+proj=utm +zone={utm_zone} +datum=WGS84 +units=m +no_defs"))
    point_utm <- st_transform(point_geom, crs = utm_crs)
    buffer_utm <- st_buffer(point_utm, dist = buffer_dist)
    buffer_wgs84 <- st_transform(buffer_utm, crs = 4326)
    exact_extract(lu_raster, buffer_wgs84, 'mean')
  }
  
  val1 <- pbapply::pbsapply(1:nrow(sf1), function(i) process_point(sf1[i, ]))
  val2 <- pbapply::pbsapply(1:nrow(sf2), function(i) process_point(sf2[i, ]))
  
  abs(val1 - val2)
}

land_types <- list(
  Build_up = "input_file/land_use/PROBAV_LC100_global_v3.0.1_2019-nrt_BuiltUp-CoverFraction-layer_EPSG-4326.tif",
  Crops    = "input_file/land_use/PROBAV_LC100_global_v3.0.1_2019-nrt_Crops-CoverFraction-layer_EPSG-4326.tif",
  Forest   = "input_file/land_use/GDLC/PROBAV_LC100_global_v3.0.1_2019-nrt_Tree-CoverFraction-layer_EPSG-4326.tif"
)

buffers <- c(500, 1000, 1500)

for (land in names(land_types)) {
  for (buf in buffers) {
    message(glue("Processing {land} with buffer {buf}m..."))
    varname <- glue("{land}_{buf}")
    data_selected[[varname]] <- compute_land_cover_diff(land_types[[land]], samples_sf1, samples_sf2, buf)
  }
}

 write_csv(data_selected, file = "output_file/data_selected_sorensen_functional_parts.csv")
```

# Optimal Land use buffer FD
```{r}
data <- read_csv("output_file/model_data_sorensen_functional_parts.csv")

response_vars <- c("turnover", "total", "nestedness")

predictor_groups <- list(
  Build_up = c("Build_up_500", "Build_up_1000", "Build_up_1500"),
  Crops    = c("Crops_500", "Crops_1000", "Crops_1500"),
  Forest   = c("Forest_500", "Forest_1000", "Forest_1500")
)

results <- data.frame()

for (response in response_vars) {
  for (group in names(predictor_groups)) {
    for (predictor in predictor_groups[[group]]) {
      formula <- as.formula(paste(response, "~", predictor))
      r2 <- tryCatch({
        model <- lm(formula, data = data)
        summary(model)$r.squared
      }, error = function(e) NA)
      results <- rbind(results, data.frame(Response = response, Group = group, Buffer = predictor, R2 = r2))
    }
  }
}

# Manually check the optimal buffer
best_results <- results %>%
  group_by(Response, Group) %>%
  slice_max(R2, n = 1, with_ties = FALSE)


print(results)
print(best_results)

```
