==> all_scripts/2.0 brm_model.Rmd <==
##########
Run brm models for taxonomic and nestedness turnover and nestednss.
Extract model results
##########

# Brm model TD
```{r}
pacman::p_load(patchwork, tidyverse, brms, tidybayes, marginaleffects)

data <- read_csv("output_file/data_selected_sorensen_parts.csv")

pairwise_vars <- c(
  "spatial.distance", "TP_diff", "Crops_1500", "Build_up_500", "Forest_1500",
  "AT", "AP", "HFP"
)

dataset_level_vars <- c("Gamma_Diversity", "Body_size")

data_turnover <- data %>%
  dplyr::select(dataset, rep, turnover, Group, Ecosystem, Dispersal_type,
                all_of(pairwise_vars), all_of(dataset_level_vars)) %>%
  dplyr::rename(
    Spatial_distance = spatial.distance,
    TP = TP_diff,
    Agriculture = Crops_1500,
    Urban = Build_up_500,
    Forest = Forest_1500,
    Temperature = AT,
    Precipitation = AP,
    Gamma_diversity = Gamma_Diversity,
    Body_size = Body_size
  )

#  log1p + scale
pairwise_vars_renamed <- c(
  "Spatial_distance", "TP", "Agriculture", "Urban", "Forest",
  "Temperature", "Precipitation", "HFP"
)

data_turnover <- data_turnover %>%
  mutate(across(all_of(pairwise_vars_renamed), log1p)) %>%
  mutate(across(all_of(pairwise_vars_renamed), scale))

dataset_scaled <- data_turnover %>%
  dplyr::select(dataset, Gamma_diversity, Body_size) %>%
  distinct() %>%
  mutate(
    Gamma_diversity = scale(log1p(Gamma_diversity)),
    Body_size = scale(log1p(Body_size))
  )

data_turnover <- data_turnover %>%
  dplyr::select(-Gamma_diversity, -Body_size) %>%
  left_join(dataset_scaled, by = "dataset")

data_turnover <- data_turnover %>%
  mutate(
    Group = relevel(factor(Group),ref ="Producer"),
    Ecosystem = relevel(factor(Ecosystem),ref ="River"),
    Dispersal_type = relevel(factor(Dispersal_type), ref = "Passive")
  )

## Fll model
fit_turn_tax <- brm(
  bf(turnover ~ Spatial_distance +
             TP * (Agriculture + Urban + Forest) +
             Group + Ecosystem + Gamma_diversity +
             Temperature + Precipitation +
             Body_size + Dispersal_type+
             (1 + Spatial_distance | dataset:rep), decomp = "QR"),
  data = data_turnover %>% mutate(rep = as.factor(rep)),
  family = zero_one_inflated_beta(),
  chains = 4,
  iter = 4000,
  warmup = 2000,
  seed = 1234,
  cores = 4,
  control = list(adapt_delta = 0.99),
  init = 0
)


saveRDS(fit_turn_tax, "model_turnover_TD.rds")



data <- read_csv("output_file/data_selected_sorensen_parts.csv")

pairwise_vars <- c(
  "spatial.distance", "TP_diff", "Crops_1500", "Build_up_500", "Forest_500",
  "AT", "AP", "HFP"
)

dataset_level_vars <- c("Gamma_Diversity", "Body_size")

data_nestedness <- data %>%
  dplyr::select(dataset, rep, nestedness, Group, Ecosystem, Dispersal_type,
                all_of(pairwise_vars), all_of(dataset_level_vars)) %>%
  dplyr::rename(
    Spatial_distance = spatial.distance,
    TP = TP_diff,
    Agriculture = Crops_1500,
    Urban = Build_up_500,
    Forest = Forest_500,
    Temperature = AT,
    Precipitation = AP,
    Gamma_diversity = Gamma_Diversity,
    Body_size = Body_size
  )

# log1p + scale
pairwise_vars_renamed <- c(
  "Spatial_distance", "TP", "Agriculture", "Urban", "Forest",
  "Temperature", "Precipitation", "HFP"
)

data_nestedness <- data_nestedness %>%
  mutate(across(all_of(pairwise_vars_renamed), log1p)) %>%
  mutate(across(all_of(pairwise_vars_renamed), scale))

dataset_scaled <- data_nestedness %>%
  dplyr::select(dataset, Gamma_diversity, Body_size) %>%
  distinct() %>%
  mutate(
    Gamma_diversity = scale(log1p(Gamma_diversity)),
    Body_size = scale(log1p(Body_size))
  )

data_nestedness <- data_nestedness %>%
  dplyr::select(-Gamma_diversity, -Body_size) %>%
  left_join(dataset_scaled, by = "dataset")

data_nestedness <- data_nestedness %>%
  mutate(
    Group = relevel(factor(Group),ref ="Producer"),
    Ecosystem = relevel(factor(Ecosystem),ref ="River"),
    Dispersal_type = relevel(factor(Dispersal_type), ref = "Passive")
  )

## Fll model
fit_nest_tax <- brm(
  bf(nestedness ~ Spatial_distance +
             TP * (Agriculture + Urban + Forest) +
             Group + Ecosystem + Gamma_diversity +
             Temperature + Precipitation +
             Body_size + Dispersal_type+
             (1 + Spatial_distance | dataset:rep), decomp = "QR"),
  data = data_nestedness %>% mutate(rep = as.factor(rep)),
  family = zero_one_inflated_beta(),
  chains = 4,
  iter = 4000,
  warmup = 2000,
  seed = 1234,
  cores = 4,
  control = list(adapt_delta = 0.99),
  init = 0
)

saveRDS(fit_nest_tax, "model_nestedness_TD2.rds")
```

# Brm model FD
```{r}
data <- read_csv("output_file/data_selected_sorensen_functional_parts.csv")

pairwise_vars <- c(
  "spatial.distance", "TP_diff", "Crops_500", "Build_up_500", "Forest_500",
  "AT", "AP", "HFP"
)

dataset_level_vars <- c("FDGamma", "Body_size")

data_turnover <- data %>%
  dplyr::select(dataset, rep, turnover, Group, Ecosystem, Dispersal_type,
                all_of(pairwise_vars), all_of(dataset_level_vars)) %>%
  dplyr::rename(
    Spatial_distance = spatial.distance,
    TP = TP_diff,
    Agriculture = Crops_500,
    Urban = Build_up_500,
    Forest = Forest_500,
    Temperature = AT,
    Precipitation = AP,
    Gamma_diversity = FDGamma,
    Body_size = Body_size
  )

# log1p + scale
pairwise_vars_renamed <- c(
  "Spatial_distance", "TP", "Agriculture", "Urban", "Forest",
  "Temperature", "Precipitation", "HFP"
)

data_turnover <- data_turnover %>%
  mutate(across(all_of(pairwise_vars_renamed), log1p)) %>%
  mutate(across(all_of(pairwise_vars_renamed), scale))

dataset_scaled <- data_turnover %>%
  dplyr::select(dataset, Gamma_diversity, Body_size) %>%
  distinct() %>%
  mutate(
    Gamma_diversity = scale(log1p(Gamma_diversity)),
    Body_size = scale(log1p(Body_size))
  )

data_turnover <- data_turnover %>%
  dplyr::select(-Gamma_diversity, -Body_size) %>%
  left_join(dataset_scaled, by = "dataset")

data_turnover <- data_turnover %>%
  mutate(
    Group = relevel(factor(Group),ref ="Producer"),
    Ecosystem = relevel(factor(Ecosystem),ref ="River"),
    Dispersal_type = relevel(factor(Dispersal_type), ref = "Passive")
  )


fit_turn_fun <- brm(
  bf(turnover ~ Spatial_distance +
             TP * (Agriculture + Urban + Forest) +
             Group + Ecosystem + Gamma_diversity +
             Temperature + Precipitation +
             Body_size + Dispersal_type+
             (1 + Spatial_distance | dataset:rep), decomp = "QR"),
  data = data_turnover %>% mutate(rep = as.factor(rep)),
  family = zero_one_inflated_beta(),
  chains = 4,
  iter = 4000,
  warmup = 2000,
  seed = 1234,
  cores = 4,
  control = list(adapt_delta = 0.99),
  init = 0
)

saveRDS(fit_turn_fun, "model_turnover_FD.rds")





data <- read_csv("output_file/data_selected_sorensen_functional_parts.csv")

pairwise_vars <- c(
  "spatial.distance", "TP_diff", "Crops_500", "Build_up_500", "Forest_500",
  "AT", "AP", "HFP"
)

dataset_level_vars <- c("FDGamma", "Body_size")

data_nestedness <- data %>%
  dplyr::select(dataset, rep, nestedness, Group, Ecosystem, Dispersal_type,
                all_of(pairwise_vars), all_of(dataset_level_vars)) %>%
  dplyr::rename(
    Spatial_distance = spatial.distance,
    TP = TP_diff,
    Agriculture = Crops_500,
    Urban = Build_up_500,
    Forest = Forest_500,
    Temperature = AT,
    Precipitation = AP,
    Gamma_diversity = FDGamma,
    Body_size = Body_size
  )

# log1p + scale
pairwise_vars_renamed <- c(
  "Spatial_distance", "TP", "Agriculture", "Urban", "Forest",
  "Temperature", "Precipitation", "HFP"
)

data_nestedness <- data_nestedness %>%
  mutate(across(all_of(pairwise_vars_renamed), log1p)) %>%
  mutate(across(all_of(pairwise_vars_renamed), scale))

dataset_scaled <- data_nestedness %>%
  dplyr::select(dataset, Gamma_diversity, Body_size) %>%
  distinct() %>%
  mutate(
    Gamma_diversity = scale(log1p(Gamma_diversity)),
    Body_size = scale(log1p(Body_size))
  )

data_nestedness <- data_nestedness %>%
  dplyr::select(-Gamma_diversity, -Body_size) %>%
  left_join(dataset_scaled, by = "dataset")

data_nestedness <- data_nestedness %>%
  mutate(
    Group = relevel(factor(Group),ref ="Producer"),
    Ecosystem = relevel(factor(Ecosystem),ref ="River"),
    Dispersal_type = relevel(factor(Dispersal_type), ref = "Passive")
  )


fit_nest_fun <- brm(
  bf(nestedness ~ Spatial_distance +
             TP * (Agriculture + Urban + Forest) +
             Group + Ecosystem + Gamma_diversity +
             Temperature + Precipitation +
             Body_size + Dispersal_type+
             (1 + Spatial_distance | dataset:rep), decomp = "QR"),
  data = data_nestedness %>% mutate(rep = as.factor(rep)),
  family = zero_one_inflated_beta(),
  chains = 4,
  iter = 4000,
  warmup = 2000,
  seed = 1234,
  cores = 4,
  control = list(adapt_delta = 0.99),
  init = 0
)

saveRDS(fit_nest_fun, "model_nestedness_FD.rds")
```



# Extract model parameters
```{r}
library(brms)
library(dplyr)
library(purrr) 
library(tibble)
## Taxonomic models
fit_turn_tax <- readRDS("model_turnover_TD.rds")
fit_nest_tax <- readRDS("model_nestedness_TD.rds")

## Functional models
fit_turn_fun <- readRDS("model_turnover_FD.rds")
fit_nest_fun <- readRDS("model_nestedness_FD.rds")

## 
summary(fit_turn_tax)
summary(fit_nest_tax)
summary(fit_turn_fun)
summary(fit_nest_fun)


## Increasing of decreasing %

target_vars <- c("Spatial_distance", "TP", "Temperature", 
                 "Precipitation", "Agriculture", "Urban", "Forest")

get_variable_impacts <- function(model, vars_list, model_name = "Model") {
  df_raw <- model$data

  result_table <- map_dfr(vars_list, function(var_name) {
    
    if (!var_name %in% names(df_raw)) {
      warning(paste("变量", var_name, "不在模型数据中，已跳过。"))
      return(NULL)
    }
    
    val_min <- min(df_raw[[var_name]], na.rm = TRUE)
    val_max <- max(df_raw[[var_name]], na.rm = TRUE)
    nd <- df_raw %>%
      select(-all_of(var_name)) %>% 
      summarise(
        across(where(is.numeric), \(x) mean(x, na.rm = TRUE)), 
        across(where(is.factor), \(x) names(sort(table(x), decreasing=TRUE))[1]),
        across(where(is.character), \(x) names(sort(table(x), decreasing=TRUE))[1])
      ) %>%
      slice(rep(1, 2)) %>% 
      mutate(!!var_name := c(val_min, val_max)) 
    post_preds <- fitted(model, newdata = nd, 
                         re_formula = NA, 
                         summary = FALSE, 
                         scale = "response")
    y_start <- post_preds[, 1] 
    y_end   <- post_preds[, 2]
    
    pct_change_dist <- (y_end - y_start) / y_start * 100
    
    tibble(
      Model = model_name,
      Variable = var_name,
      Min_Value = val_min,
      Max_Value = val_max,
      Pct_Change = median(pct_change_dist),
      CI_Low = quantile(pct_change_dist, 0.025),
      CI_High = quantile(pct_change_dist, 0.975)
    )
  })
  
  return(result_table)
}

# Taxonomic Turnover
res_turn_tax <- get_variable_impacts(fit_turn_tax, target_vars, "Taxonomic Turnover")

# axonomic Nestedness
res_nest_tax <- get_variable_impacts(fit_nest_tax, target_vars, "Taxonomic Nestedness")

# Functional Turnover
res_turn_fun <- get_variable_impacts(fit_turn_fun, target_vars, "Functional Turnover")

# Functional Nestedness
res_nest_fun <- get_variable_impacts(fit_nest_fun, target_vars, "Functional Nestedness")


final_results <- bind_rows(res_turn_tax, res_nest_tax, res_turn_fun, res_nest_fun)

print(final_results)


```
