==> all_scripts/3.0 plot_figure_2.Rmd <==
##########
Plot site map and bar charts of variables
Combine them into Fig.2
##########


# Fig 2. Site map
```{r}
library(sf)
library(ggplot2)
library(dplyr)
library(rnaturalearth)
library(rnaturalearthdata)

dat <- read.csv("input_file/dataset_var.csv")

dat_rob <- dat %>%
  filter(!is.na(X), !is.na(Y)) %>%
  st_as_sf(coords = c("X", "Y"), crs = 4326) %>%
  st_transform(crs = "+proj=robin +lon_0=0 +datum=WGS84 +units=m +no_defs")

world_map <- ne_countries(scale = "medium", returnclass = "sf") %>%
  filter(continent != "Antarctica") %>%
  st_transform(crs = "+proj=robin +lon_0=0 +datum=WGS84 +units=m +no_defs")

color_palette <- c("Producer" = "#1f77b4", "Consumer" = "#ff7f0e")
shape_palette <- c("Lake" = 16 , "River" = 1)  

size_breaks <- c(50, 100, 500, 1500, 2000)

scale_size_continuous(
  name = "Site Number",
  breaks = size_breaks,
  labels = size_breaks,
  range = c(1.5, 6)  
)

ggplot() +
  geom_sf(data = world_map, fill = "grey90", color = "grey80", linewidth = 0.2, alpha = 0.8) +
  geom_sf(
    data = dat_rob,
    aes(color = Group, shape = Ecosystem, size = Site_num),
    alpha = 0.8
  ) +
  scale_color_manual(values = color_palette, name = "Group") +
  scale_shape_manual(values = shape_palette, name = "Ecosystem") +
  scale_size_continuous(
    name = "Site Number",
    breaks = c(50, 100, 500, 1500, 2000),
    labels = c("50", "100", "500", "1500", "2000"),
    range = c(1.5, 6)
  ) +
  guides(
    size  = guide_legend(order = 1, override.aes = list(shape = 16)),
    shape = guide_legend(order = 2),
    color = guide_legend(order = 3)
  ) +
  theme_void() +
  theme(
    legend.position = "bottom",              
    legend.box = "horizontal",              
    legend.direction = "horizontal",         
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 9),
    legend.spacing.x = unit(0.8, "cm"),
    legend.spacing.y = unit(0.3, "cm"),
    plot.margin = margin(0, 0, 0, 0)
  )

#ggsave("output_figure/Fig.2-1.pdf", width = 10, height = 6, dpi = 300)

```

# Fig 2. Bar chart
```{r}
dat <- read.csv("input_file/dataset_var.csv")

dat_log <- dat %>%
  filter(Ave_TP > 0, !is.na(Ave_TP), !is.na(Ave_AT), !is.na(Ave_AP)) %>%
  mutate(
    logTP = log(Ave_TP)
  )

custom_theme <- theme_minimal() +
  theme(
    panel.border = element_blank(),
    axis.line.x = element_line(color = "black"),
    axis.line.y = element_line(color = "black"),
    panel.grid = element_blank(),
    axis.ticks = element_line(color = "black"),
    axis.text = element_text(color = "black")
  )

p1 <- ggplot(dat_log, aes(x = logTP)) +
  geom_histogram(bins = 12, fill = "steelblue3", color = "white") +
  labs(x = "Total phosphorus (log)", y = "Number of Datasets") +
  custom_theme

# Ave_AT
p2 <- ggplot(dat_log, aes(x = Ave_AT)) +
  geom_histogram(bins = 12, fill = "steelblue3", color = "white") +
  labs(x = "Annual Temperature (°C)", y = "Number of Datasets") +
  custom_theme

# Ave_AP
p3 <- ggplot(dat_log, aes(x = Ave_AP)) +
  geom_histogram(bins = 12, fill = "steelblue3", color = "white") +
  labs(x = "Precipitation (mm)", y = "Number of Datasets") +
  custom_theme

# Gamma
p4 <- ggplot(dat_log, aes(x = Gamma_Diversity)) +
  geom_histogram(bins = 12, fill = "steelblue3", color = "white") +
  labs(x = "Gamma Diversity", y = "Number of Datasets") +
  custom_theme

library(gridExtra)
p_total=grid.arrange(p1, p2, p3,p4, ncol = 4)
#ggsave("output_figure/Fig.2-3.pdf", p_total, width = 10, height = 2.5, dpi = 300)



eco_summary <- dat %>%
  count(Category = Ecosystem) %>%
  mutate(Type = "Ecosystem")

grp_summary <- dat %>%
  count(Category = Group) %>%
  mutate(Type = "Group")

disp_summary <- dat %>%
  count(Category = Dispersal_type) %>%
  mutate(Type = "Dispersal_type")

combined_summary <- bind_rows(eco_summary, grp_summary, disp_summary)
combined_summary$Category <- factor(
  combined_summary$Category,
  levels = rev(c("River", "Lake", "Producer", "Consumer", "Passive", "Active", "Seeds"))
)

ggplot(combined_summary, aes(x = n, y = Category, fill = Type)) +
  geom_bar(stat = "identity") +
  scale_fill_manual(values = c("Ecosystem" = "#547ac0", 
                               "Group" = "steelblue3", 
                               "Dispersal_type" = "#99cbeb")) +
  labs(
    x = "Number of Datasets",
    y = NULL,
    fill = NULL
  ) +
  theme_minimal() +
  theme(
    panel.grid = element_blank(),
    axis.line.x = element_line(color = "black"),
    axis.line.y = element_line(color = "black"),
    axis.ticks = element_line(color = "black"),
    axis.text = element_text(color = "black"),
    legend.position = "none"
  )

#ggsave("output_figure/Fig.2-2.pdf", width = 3, height = 3, dpi = 300)
```
