==> all_scripts/3.1 plot_figure_3.Rmd <==
##########
Plot Fig.3
##########

# Figure 3 TD
```{r}
library(tidyverse)
library(modelr)
library(tidybayes)
library(marginaleffects)
library(patchwork)
library(brms)
library(ggdist)
library(ggplot2)

fit_turn <- readRDS("output_file/model_turnover_TD.rds")
fit_nest <- readRDS("output_file/model_nestedness_TD.rds")

desired_order <- c("Turnover", "Nestedness")

# TP
tp_seq <- modelr::seq_range(fit_turn$data$TP, n = 100)
newdata_tp <- datagrid(model = fit_turn, TP = tp_seq)

# Agriculture
agri_seq <- modelr::seq_range(fit_turn$data$Agriculture, n = 100)
newdata_agri <- datagrid(model = fit_turn, Agriculture = agri_seq)

# Urban
urban_seq <- modelr::seq_range(fit_turn$data$Urban, n = 100)
newdata_urban <- datagrid(model = fit_turn, Urban = urban_seq)

# Forest
forest_seq <- modelr::seq_range(fit_turn$data$Forest, n = 100)
newdata_forest <- datagrid(model = fit_turn, Forest = forest_seq)

# Ecosystem & Group
eco_levels <- unique(fit_turn$data$Ecosystem)
grp_levels <- unique(fit_turn$data$Group)
newdata_eco <- datagrid(model = fit_turn, Ecosystem = eco_levels)
newdata_grp <- datagrid(model = fit_turn, Group = grp_levels)


get_preds <- function(newdata) {
  p_turn <- epred_draws(fit_turn, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Turnover", levels = desired_order))
  p_nest <- epred_draws(fit_nest, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Nestedness", levels = desired_order))
  bind_rows(p_turn, p_nest)
}

all_preds_tp     <- get_preds(newdata_tp)
all_preds_agri   <- get_preds(newdata_agri)
all_preds_urban  <- get_preds(newdata_urban)
all_preds_forest <- get_preds(newdata_forest)
all_preds_eco    <- get_preds(newdata_eco) %>% mutate(X = as.character(Ecosystem))
all_preds_grp    <- get_preds(newdata_grp) %>% mutate(X = as.character(Group))


tp_center <- attr(fit_turn$data$TP, "scaled:center")
tp_scale  <- attr(fit_turn$data$TP, "scaled:scale")
all_preds_tp <- all_preds_tp %>% 
  mutate(TP_orig = TP * tp_scale + tp_center)

agri_center <- attr(fit_turn$data$Agriculture, "scaled:center")
agri_scale  <- attr(fit_turn$data$Agriculture, "scaled:scale")
all_preds_agri <- all_preds_agri %>% 
  mutate(Agriculture_orig = expm1(Agriculture * agri_scale + agri_center))

urban_center <- attr(fit_turn$data$Urban, "scaled:center")
urban_scale  <- attr(fit_turn$data$Urban, "scaled:scale")
all_preds_urban <- all_preds_urban %>% 
  mutate(Urban_orig = expm1(Urban * urban_scale + urban_center))

forest_center <- attr(fit_turn$data$Forest, "scaled:center")
forest_scale  <- attr(fit_turn$data$Forest, "scaled:scale")
all_preds_forest <- all_preds_forest %>% 
  mutate(Forest_orig = expm1(Forest * forest_scale + forest_center))

fill_colors <- c("Turnover" = "#FFA500", "Nestedness" = "#3182BD")
line_colors <- c("Turnover" = "#A0522D", "Nestedness" = "#08519C")

# --- 1. Total Phosphorus ---
fig_tp <- ggplot(all_preds_tp, aes(x = TP_orig, y = .epred, fill = .category, color = .category)) + 
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, 
                  linewidth = 0.4) + 
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = expression(TP~"("*mu*g~L^{-1}*", log)"), 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 2. Agriculture ---
fig_agri <- ggplot(all_preds_agri, aes(x = Agriculture_orig, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, 
                  linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = "Agriculture cover (%)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 3. Urban ---
fig_urban <- ggplot(all_preds_urban, aes(x = Urban_orig, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, 
                  linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  # 单位：Urban cover (%)
  labs(x = "Urban cover (%)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 4. Forest ---
fig_forest <- ggplot(all_preds_forest, aes(x = Forest_orig, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, 
                  linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  # 单位：Forest cover (%)
  labs(x = "Forest cover (%)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 5. Ecosystem (Categorical) ---
fig_eco <- ggplot(all_preds_eco, aes(x = X, y = .epred, colour = .category, group = .category)) +
  stat_interval(aes(colour_ramp = after_stat(level)), 
                position = position_dodge(width = 0.6),
                .width = c(0.8, 0.9, 0.95), 
                point_interval = "median_qi", 
                show.legend = FALSE) +
  scale_colour_manual(values = fill_colors, guide = "none") +
  labs(y = "Taxonomic β-diversity", x = NULL) + 
  theme_bw() +
  theme(axis.text.x = element_text(size = 10), 
        axis.title.y = element_text(size = 10), 
        panel.grid = element_blank())

# --- 6. Group (Categorical) ---
fig_grp <- ggplot(all_preds_grp, aes(x = X, y = .epred, colour = .category, group = .category)) +
  stat_interval(aes(colour_ramp = after_stat(level)), 
                position = position_dodge(width = 0.6),
                .width = c(0.8, 0.9, 0.95), 
                point_interval = "median_qi", 
                show.legend = FALSE) +
  scale_x_discrete(limits = c("Producer", "Consumer")) +
  scale_colour_manual(values = fill_colors, guide = "none") +
  labs(y = "Taxonomic β-diversity", x = NULL) + 
  theme_bw() +
  theme(axis.text.x = element_text(size = 10), 
        axis.title.y = element_text(size = 10), 
        panel.grid = element_blank())

fig_TD <- (fig_tp + fig_urban + fig_eco + fig_agri + fig_forest + fig_grp) +
  plot_layout(ncol = 3, nrow = 2, guides = "collect") +
  plot_annotation( theme = theme(legend.position = "bottom"))

fig_TD

```

# Figure 3 FD
```{r}
fit_turn <- readRDS("output_file/model_turnover_FD.rds")
fit_nest <- readRDS("output_file/model_nestedness_FD.rds")

desired_order <- c("Turnover", "Nestedness")

# TP
tp_seq <- modelr::seq_range(fit_turn$data$TP, n = 100)
newdata_tp <- datagrid(model = fit_turn, TP = tp_seq)

# Agriculture
agri_seq <- modelr::seq_range(fit_turn$data$Agriculture, n = 100)
newdata_agri <- datagrid(model = fit_turn, Agriculture = agri_seq)

# Urban
urban_seq <- modelr::seq_range(fit_turn$data$Urban, n = 100)
newdata_urban <- datagrid(model = fit_turn, Urban = urban_seq)

# Forest
forest_seq <- modelr::seq_range(fit_turn$data$Forest, n = 100)
newdata_forest <- datagrid(model = fit_turn, Forest = forest_seq)

# Ecosystem & Group
eco_levels <- unique(fit_turn$data$Ecosystem)
grp_levels <- unique(fit_turn$data$Group)
newdata_eco <- datagrid(model = fit_turn, Ecosystem = eco_levels)
newdata_grp <- datagrid(model = fit_turn, Group = grp_levels)

get_preds <- function(newdata) {
  p_turn <- epred_draws(fit_turn, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Turnover", levels = desired_order))
  p_nest <- epred_draws(fit_nest, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Nestedness", levels = desired_order))
  bind_rows(p_turn, p_nest)
}

all_preds_tp     <- get_preds(newdata_tp)
all_preds_agri   <- get_preds(newdata_agri)
all_preds_urban  <- get_preds(newdata_urban)
all_preds_forest <- get_preds(newdata_forest)
all_preds_eco    <- get_preds(newdata_eco) %>% mutate(X = as.character(Ecosystem))
all_preds_grp    <- get_preds(newdata_grp) %>% mutate(X = as.character(Group))

tp_center <- attr(fit_turn$data$TP, "scaled:center")
tp_scale  <- attr(fit_turn$data$TP, "scaled:scale")
all_preds_tp <- all_preds_tp %>% 
  mutate(TP_orig = TP * tp_scale + tp_center)

agri_center <- attr(fit_turn$data$Agriculture, "scaled:center")
agri_scale  <- attr(fit_turn$data$Agriculture, "scaled:scale")
all_preds_agri <- all_preds_agri %>% 
  mutate(Agriculture_orig = expm1(Agriculture * agri_scale + agri_center))

urban_center <- attr(fit_turn$data$Urban, "scaled:center")
urban_scale  <- attr(fit_turn$data$Urban, "scaled:scale")
all_preds_urban <- all_preds_urban %>% 
  mutate(Urban_orig = expm1(Urban * urban_scale + urban_center))

forest_center <- attr(fit_turn$data$Forest, "scaled:center")
forest_scale  <- attr(fit_turn$data$Forest, "scaled:scale")
all_preds_forest <- all_preds_forest %>% 
  mutate(Forest_orig = expm1(Forest * forest_scale + forest_center))

fill_colors <- c("Turnover" = "#FFA500", "Nestedness" = "#3182BD")
line_colors <- c("Turnover" = "#A0522D", "Nestedness" = "#08519C")

# --- 1. Total Phosphorus ---
fig_tp <- ggplot(all_preds_tp, aes(x = TP_orig, y = .epred, fill = .category, color = .category)) + 
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) + 
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::number_format(accuracy = 0.1)) +
  labs(x = expression(TP~"("*mu*g~L^{-1}*", log)"), 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 2. Agriculture ---
fig_agri <- ggplot(all_preds_agri, aes(x = Agriculture_orig, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  
  scale_y_continuous(
    breaks = seq(0, 1, 0.1), 
    labels = scales::number_format(accuracy = 0.1),
    limits = c(0.1, NA) 
  ) +
  
  labs(x = "Agriculture cover (%)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 3. Urban ---
fig_urban <- ggplot(all_preds_urban, aes(x = Urban_orig, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::number_format(accuracy = 0.1)) +
  labs(x = "Urban cover (%)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 4. Forest ---
fig_forest <- ggplot(all_preds_forest, aes(x = Forest_orig, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), 
                  point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::number_format(accuracy = 0.1),
    limits = c(0.1, NA) ) +
  labs(x = "Forest cover (%)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", 
       fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# --- 5. Ecosystem (Categorical) ---
fig_eco <- ggplot(all_preds_eco, aes(x = X, y = .epred, colour = .category, group = .category)) +
  stat_interval(aes(colour_ramp = after_stat(level)), 
                position = position_dodge(width = 0.6),
                .width = c(0.8, 0.9, 0.95), 
                point_interval = "median_qi", 
                show.legend = FALSE) +
  scale_colour_manual(values = fill_colors, guide = "none") +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::number_format(accuracy = 0.1)) +
  labs(y = "Taxonomic β-diversity", x = NULL) + 
  theme_bw() +
  theme(axis.text.x = element_text(size = 10), 
        axis.title.y = element_text(size = 10), 
        panel.grid = element_blank())

# --- 6. Group (Categorical) ---
fig_grp <- ggplot(all_preds_grp, aes(x = X, y = .epred, colour = .category, group = .category)) +
  stat_interval(aes(colour_ramp = after_stat(level)), 
                position = position_dodge(width = 0.6),
                .width = c(0.8, 0.9, 0.95), 
                point_interval = "median_qi", 
                show.legend = FALSE) +
  scale_x_discrete(limits = c("Producer", "Consumer")) +
  scale_colour_manual(values = fill_colors, guide = "none") +
  scale_y_continuous(breaks = seq(0, 1, 0.1), labels = scales::number_format(accuracy = 0.1)) +
  labs(y = "Taxonomic β-diversity", x = NULL) + 
  theme_bw() +
  theme(axis.text.x = element_text(size = 10), 
        axis.title.y = element_text(size = 10), 
        panel.grid = element_blank())

fig_FD <- (fig_tp + fig_urban + fig_eco + fig_agri + fig_forest + fig_grp) +
  plot_layout(ncol = 3, nrow = 2, guides = "collect") +
  plot_annotation(
    theme = theme(legend.position = "bottom")
  )

fig_all <- (fig_TD / fig_FD) +
  plot_layout(guides = "collect") +
  plot_annotation(theme = theme(legend.position = "bottom"))

fig_all

ggsave("output_figure/Fig.3.pdf", fig_all, height = 11, width = 7, unit = "in")
```
