==> all_scripts/3.2 plot_figure_4.Rmd <==
##########
Plot Fig.4
##########


# Figure 4
```{r}
library(patchwork) 
library(tidyverse)
library(brms)       
library(broom)
library(broom.mixed)


extract_params <- function(model, component_label, metric_label) {
  fixef(model) %>%
    as.data.frame() %>%
    rownames_to_column("parameter") %>%
    as_tibble() %>%
    filter(parameter != "Intercept" & parameter != "b_Intercept") %>% 
    mutate(
      predictor = str_remove(parameter, "^b_"), 
      Component = component_label,
      Metric    = metric_label
    )
}

fit_turn_td <- readRDS("model_turnover_TD.rds")
fit_nest_td <- readRDS("model_nestedness_TD.rds")

params_turn_td <- extract_params(fit_turn_td, "Turnover", "Taxonomic")
params_nest_td <- extract_params(fit_nest_td, "Nestedness", "Taxonomic")

fit_turn_fd <- readRDS("model_turnover_FD.rds")
fit_nest_fd <- readRDS("model_nestedness_FD.rds")

params_turn_fd <- extract_params(fit_turn_fd, "Turnover", "Functional")
params_nest_fd <- extract_params(fit_nest_fd, "Nestedness", "Functional")


params_final <- bind_rows(params_turn_td, params_nest_td, params_turn_fd, params_nest_fd)


pairwise_vars <- c("Spatial_distance", "TP","Agriculture","Urban","Forest",
                   "TP:Agriculture","TP:Urban","TP:Forest",
                   "Temperature","Precipitation")
dataset_vars  <- c("EcosystemLake","GroupConsumer","Gamma_diversity",
                   "Body_size","Dispersal_typeActive","Dispersal_typeSeeds")

label_map <- c(
  "GroupConsumer"        = "Group: Consumer",
  "EcosystemLake"        = "Ecosystem: Lake",
  "Gamma_diversity"      = "Gamma diversity",
  "Body_size"            = "Body size",
  "Dispersal_typeActive" = "Dispersal type: Active",
  "Dispersal_typeSeeds"  = "Dispersal type: Seeds",
  "Spatial_distance"     = "Spatial distance",
  "TP"                   = "TP",
  "Agriculture"          = "Agriculture",
  "Urban"                = "Urban",
  "Forest"               = "Forest",
  "Temperature"          = "Temperature",
  "Precipitation"        = "Precipitation",
  "TP:Agriculture"       = "TP*Agriculture",
  "TP:Urban"             = "TP*Urban",
  "TP:Forest"            = "TP*Forest"
)

label_order <- c(label_map[pairwise_vars], label_map[dataset_vars]) %>% as.vector()

params_plot <- params_final %>%
  mutate(
    is_sig       = !(Q2.5 < 0 & Q97.5 > 0), 
    Significance = ifelse(is_sig, "Sig", "Insig"),
    Significance = factor(Significance, levels = c("Sig", "Insig")), 
    Component    = factor(Component, levels = c("Turnover", "Nestedness")),
    Metric       = factor(Metric, levels = c("Taxonomic", "Functional")),
    Variable_type   = if_else(predictor %in% pairwise_vars, "Site-level variables", "Dataset-level variables"),
    predictor_label = dplyr::recode(predictor, !!!label_map, .default = predictor),
    predictor_label = factor(predictor_label, levels = rev(label_order))
  )

common_theme <- theme_bw(base_size = 12) +
  theme(
    panel.grid.major  = element_blank(),
    panel.grid.minor  = element_blank(),
    
    strip.background.x = element_rect(fill = "grey90", color = "grey50", linewidth = 1),
    
    strip.background.y = element_rect(fill = "white", color = NA),
    
    strip.text        = element_text(face = "bold", size = 11),
    strip.placement   = "outside", 
    
    legend.position   = "bottom",
    legend.title      = element_text(),
    axis.text.y       = element_text(color = "black", size = 10),
    plot.title        = element_blank() 
  )

metric_colors <- c("Taxonomic" = "#1f77b4", "Functional" = "#db6968")
shape_values  <- c("Sig" = 16, "Insig" = 21) 
pd <- position_dodge(width = 0.6)

data_upper <- params_plot %>% filter(Variable_type == "Site-level variables")

p_upper <- ggplot(data_upper, 
                  aes(y = predictor_label, x = Estimate, color = Metric, group = Metric)) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey40") +
  geom_errorbarh(aes(xmin = Q2.5, xmax = Q97.5), height = 0, linewidth = 0.8, position = pd) +
  geom_point(aes(shape = Significance), size = 3, stroke = 1, position = pd, fill = "white") +
  scale_color_manual(values = metric_colors) +
  scale_shape_manual(values = shape_values) +
  
  labs(x = NULL, y = NULL, color = NULL, shape = NULL) + 
  
  facet_grid(Variable_type ~ Component, scales = "free_x", switch = "y") +
  common_theme +
  theme(plot.margin = margin(b = 5))

data_lower <- params_plot %>% filter(Variable_type == "Dataset-level variables")

p_lower <- ggplot(data_lower, 
                  aes(y = predictor_label, x = Estimate, color = Metric, group = Metric)) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey40") +
  geom_errorbarh(aes(xmin = Q2.5, xmax = Q97.5), height = 0, linewidth = 0.8, position = pd) +
  geom_point(aes(shape = Significance), size = 3, stroke = 1, position = pd, fill = "white") +
  scale_color_manual(values = metric_colors) +
  scale_shape_manual(values = shape_values) +
  
  labs(x = "Estimate", y = NULL, color = NULL, shape = NULL) +
  
  facet_grid(Variable_type ~ Component, scales = "free_x", switch = "y") +
  common_theme +
  theme(plot.margin = margin(t = 2))

n_up <- length(unique(data_upper$predictor))
n_lo <- length(unique(data_lower$predictor))

Final_Plot <- (p_upper / p_lower) + 
  plot_layout(guides = "collect", heights = c(n_up, n_lo)) & 
  theme(legend.position = "bottom",
        legend.box = "horizontal") 

print(Final_Plot)


ggsave("out_put_figure/Fig.4.pdf", 
       Final_Plot, height = 6.5, width = 6.5, unit = "in")

```
