==> all_scripts/3.3 plot_figure_S1-S3.Rmd <==
##########
Plot Fig.S1 Fig.S2 Fig.S3
##########

# Fig.S1
```{r}
library(ggplot2)
library(sf)
library(dplyr)
library(lwgeom) 

sf::sf_use_s2(FALSE)
robin_crs <- "+proj=robin +lon_0=0 +datum=WGS84 +units=m +no_defs"

dat <- read.csv("output_file/all_data_plot.csv")

dat_rob <- dat %>%
  filter(!is.na(X), !is.na(Y)) %>%
  st_as_sf(coords = c("X", "Y"), crs = 4326) %>%
  st_transform(crs = st_crs(robin_crs)) %>%
  slice_sample(prop = 1) 

organism_cols <- c(
  "Aquatic Plants"            = "#08519C",
  "Fish"                      = "#fc975f",
  "Benthic Diatoms"           = "#1f77b4",
  "Phytoplankton"             = "#3C9BC9",
  "Benthic Macroinvertebrates"= "#ff7f0e",
  "Zooplankton"               = "#d95f02"
  
)
dat_rob$Organism <- factor(dat_rob$Organism, levels = names(organism_cols))

shp_path <- "/input_file/world_shp/TM_WORLD_BORDERS_SIMPL-0.3.shp"

world_map <- st_read(shp_path, quiet = TRUE) %>%
  st_cast("POLYGON") %>%
  mutate(area = st_area(geometry)) %>%  
  filter(as.numeric(area) > 10^9) %>%   
  st_crop(st_bbox(c(xmin = -180, xmax = 180, ymin = -90, ymax = 90), crs = 4326)) %>%
  st_transform(crs = st_crs(robin_crs))

p <- ggplot() +
  geom_sf(
    data  = world_map,
    fill  = "grey90",
    color = "grey80",
    linewidth = 0.2
  ) +
  geom_sf(
    data = dat_rob,
    aes(color = Organism),
    shape = 16,    
    size  = 0.6,   
    alpha = 0.6
  ) +
  scale_color_manual(values = organism_cols, name = NULL) +
  scale_x_continuous(breaks = seq(-180, 180, 45)) +
  
  coord_sf(
    crs = st_crs(robin_crs),
    xlim = c(-17000000, 17000000), 
    ylim = c(-8600000, 8600000),
    expand = TRUE
  ) +
  
  theme_minimal() +
  theme(
    axis.text = element_blank(),
    axis.title = element_blank(),
    panel.grid.major = element_line(color = "#E0E0E0", linewidth = 0.2), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.title    = element_blank(),
    legend.text     = element_text(size = 9),
    legend.key      = element_blank(),
    plot.margin     = margin(0, 0, 0, 0)
  )

print(p)

#ggsave("output_figure/Fig.S1.pdf", p, height = 5, width = 4.5, unit = "in")
```

# Fig.S2 TD
```{r}
library(tidyverse)
library(modelr)
library(tidybayes)
library(marginaleffects)
library(patchwork)
library(brms)
library(ggdist)
library(ggplot2)

fit_turn <- readRDS("output_file/model_turnover_TD.rds")
fit_nest <- readRDS("output_file/model_nestedness_TD.rds")

desired_order <- c("Turnover", "Nestedness")
fill_colors <- c("Turnover" = "#FFA500", "Nestedness" = "#3182BD")
line_colors <- c("Turnover" = "#A0522D", "Nestedness" = "#08519C")

get_preds <- function(newdata) {
  p_turn <- epred_draws(fit_turn, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Turnover", levels = desired_order))
  p_nest <- epred_draws(fit_nest, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Nestedness", levels = desired_order))
  bind_rows(p_turn, p_nest)
}

# Spatial Distance
dist_seq <- modelr::seq_range(fit_turn$data$Spatial_distance, n = 100)
newdata_dist <- datagrid(model = fit_turn, Spatial_distance = dist_seq)
preds_dist <- get_preds(newdata_dist)

dist_center <- attr(fit_turn$data$Spatial_distance, "scaled:center")
dist_scale  <- attr(fit_turn$data$Spatial_distance, "scaled:scale")
preds_dist <- preds_dist %>% 
  mutate(Dist_log = Spatial_distance * dist_scale + dist_center)

fig_dist <- ggplot(preds_dist, aes(x = Dist_log, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  # 标签：(km, log)
  labs(x = "Spatial distance (km, log)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

#Temperature
temp_seq <- modelr::seq_range(fit_turn$data$Temperature, n = 100)
newdata_temp <- datagrid(model = fit_turn, Temperature = temp_seq)
preds_temp <- get_preds(newdata_temp)

temp_center <- attr(fit_turn$data$Temperature, "scaled:center")
temp_scale  <- attr(fit_turn$data$Temperature, "scaled:scale")
preds_temp <- preds_temp %>% 
  mutate(Temp_log = Temperature * temp_scale + temp_center)

fig_temp <- ggplot(preds_temp, aes(x = Temp_log, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = "Temperature (°C, log)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# Precipitation
precip_seq <- modelr::seq_range(fit_turn$data$Precipitation, n = 100)
newdata_precip <- datagrid(model = fit_turn, Precipitation = precip_seq)
preds_precip <- get_preds(newdata_precip)

precip_center <- attr(fit_turn$data$Precipitation, "scaled:center")
precip_scale  <- attr(fit_turn$data$Precipitation, "scaled:scale")
preds_precip <- preds_precip %>% 
  mutate(Precip_log = Precipitation * precip_scale + precip_center)

fig_precip <- ggplot(preds_precip, aes(x = Precip_log, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = "Precipitation (mm, log)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# Dispersal Type
disp_levels <- unique(fit_turn$data$Dispersal_type)
newdata_disp <- datagrid(model = fit_turn, Dispersal_type = disp_levels)
preds_disp <- get_preds(newdata_disp) %>% 
  mutate(X = as.character(Dispersal_type))

fig_disp <- ggplot(preds_disp, aes(x = X, y = .epred, colour = .category, group = .category)) +
  stat_interval(aes(colour_ramp = after_stat(level)), 
                position = position_dodge(width = 0.6),
                .width = c(0.8, 0.9, 0.95), point_interval = "median_qi", 
                show.legend = FALSE) +
  scale_colour_manual(values = fill_colors, guide = "none") +
  labs(y = "Taxonomic β-diversity", x = NULL) + 
  theme_bw() +
  theme(axis.text.x = element_text(size = 10), 
        axis.title.y = element_text(size = 10), 
        panel.grid = element_blank())

fig_TD <- (fig_dist + fig_temp + fig_precip + fig_disp) +
  plot_layout(ncol = 2, nrow = 2, guides = "collect") +
  plot_annotation(tag_levels = 'A', theme = theme(legend.position = "bottom"))

#fig_TD
```

# Fig.S2 FD
```{r}
fit_turn <- readRDS("output_file/model_turnover_FD.rds")
fit_nest <- readRDS("output_file/model_nestedness_FD.rds")

desired_order <- c("Turnover", "Nestedness")
fill_colors <- c("Turnover" = "#FFA500", "Nestedness" = "#3182BD")
line_colors <- c("Turnover" = "#A0522D", "Nestedness" = "#08519C")

get_preds <- function(newdata) {
  p_turn <- epred_draws(fit_turn, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Turnover", levels = desired_order))
  p_nest <- epred_draws(fit_nest, newdata = newdata, re_formula = NA, ndraws = 1000) %>%
    mutate(.category = factor("Nestedness", levels = desired_order))
  bind_rows(p_turn, p_nest)
}

#Spatial Distance
dist_seq <- modelr::seq_range(fit_turn$data$Spatial_distance, n = 100)
newdata_dist <- datagrid(model = fit_turn, Spatial_distance = dist_seq)
preds_dist <- get_preds(newdata_dist)

dist_center <- attr(fit_turn$data$Spatial_distance, "scaled:center")
dist_scale  <- attr(fit_turn$data$Spatial_distance, "scaled:scale")
preds_dist <- preds_dist %>% 
  mutate(Dist_log = Spatial_distance * dist_scale + dist_center)

fig_dist <- ggplot(preds_dist, aes(x = Dist_log, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = "Spatial distance (km, log)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# Temperature
temp_seq <- modelr::seq_range(fit_turn$data$Temperature, n = 100)
newdata_temp <- datagrid(model = fit_turn, Temperature = temp_seq)
preds_temp <- get_preds(newdata_temp)

temp_center <- attr(fit_turn$data$Temperature, "scaled:center")
temp_scale  <- attr(fit_turn$data$Temperature, "scaled:scale")
preds_temp <- preds_temp %>% 
  mutate(Temp_log = Temperature * temp_scale + temp_center)

fig_temp <- ggplot(preds_temp, aes(x = Temp_log, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = "Temperature (°C, log)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# Precipitation
precip_seq <- modelr::seq_range(fit_turn$data$Precipitation, n = 100)
newdata_precip <- datagrid(model = fit_turn, Precipitation = precip_seq)
preds_precip <- get_preds(newdata_precip)

precip_center <- attr(fit_turn$data$Precipitation, "scaled:center")
precip_scale  <- attr(fit_turn$data$Precipitation, "scaled:scale")
preds_precip <- preds_precip %>% 
  mutate(Precip_log = Precipitation * precip_scale + precip_center)

fig_precip <- ggplot(preds_precip, aes(x = Precip_log, y = .epred, fill = .category, color = .category)) +
  stat_lineribbon(aes(fill_ramp = after_stat(level)), 
                  .width = c(0.8, 0.9, 0.95), point_interval = "median_hdci", 
                  alpha = 0.7, linewidth = 0.4) +
  scale_fill_manual(values = fill_colors, breaks = desired_order) +
  scale_color_manual(values = line_colors, breaks = desired_order) +
  labs(x = "Precipitation (mm, log)", 
       y = "Taxonomic β-diversity",
       fill = "Component", color = "Component", fill_ramp = "Credible interval") +
  theme_bw() +
  theme(axis.title = element_text(size = 10), panel.grid = element_blank())

# Dispersal Type
disp_levels <- unique(fit_turn$data$Dispersal_type)
newdata_disp <- datagrid(model = fit_turn, Dispersal_type = disp_levels)
preds_disp <- get_preds(newdata_disp) %>% 
  mutate(X = as.character(Dispersal_type))

fig_disp <- ggplot(preds_disp, aes(x = X, y = .epred, colour = .category, group = .category)) +
  stat_interval(aes(colour_ramp = after_stat(level)), 
                position = position_dodge(width = 0.6),
                .width = c(0.8, 0.9, 0.95), point_interval = "median_qi", 
                show.legend = FALSE) +
  scale_colour_manual(values = fill_colors, guide = "none") +
  labs(y = "Taxonomic β-diversity", x = NULL) + 
  theme_bw() +
  theme(axis.text.x = element_text(size = 10), 
        axis.title.y = element_text(size = 10), 
        panel.grid = element_blank())

fig_FD <- (fig_dist + fig_temp + fig_precip + fig_disp) +
  plot_layout(ncol = 2, nrow = 2, guides = "collect") +
  plot_annotation(tag_levels = 'A', theme = theme(legend.position = "bottom"))


fig_all <- (fig_TD / fig_FD) +
  plot_layout(guides = "collect") +
  plot_annotation(theme = theme(legend.position = "bottom"))

#fig_all

ggsave("output_figure/Fig.S2.pdf", fig_all, height = 12, width = 5, unit = "in")

```

# Fig.S3
```{r}
library(ggplot2)
library(patchwork)
data_with_lu = read.csv("output_file/all_data_plot.csv")
data_plot <- data_with_lu[!is.na(data_with_lu$TP) & data_with_lu$TP > 0, ]
data_plot$ln_TP <- log1p(data_plot$TP)

make_scatter_lu_tp <- function(df, x_var, x_lab,
                               y_lab = "TP (µg L⁻¹,log-transformed)") {

  df_sub <- df[, c("ln_TP", x_var)]
  df_sub <- df_sub[complete.cases(df_sub), ]
  
  form <- as.formula(paste("ln_TP ~", x_var))
  fit  <- lm(form, data = df_sub)
  summ <- summary(fit)
  
  slope  <- coef(fit)[2]
  r2_val <- summ$r.squared
  p_val  <- coef(summ)[2, "Pr(>|t|)"]
  
  stars <- if (p_val < 0.001) {
    "***"
  } else if (p_val < 0.01) {
    "**"
  } else if (p_val < 0.05) {
    "*"
  } else {
    ""
  }
  
  label_txt <- paste0(
    "Slope = ", signif(slope, 3),
    ", R² = ", round(r2_val, 2),
    stars
  )
  
  ggplot(df_sub, aes_string(x = x_var, y = "ln_TP")) +
    geom_point(
      color = "#4a68c4",  
      alpha = 0.4,       
      size  = 1.2         
    ) +                 
    # ============================================
    geom_smooth(
      method = "lm",
      se = TRUE,                
      level = 0.95,             
      linewidth = 0.6,
      colour = "#FFA500",      
      fill   = "#FDCD94",       
      alpha  = 0.4              
    ) +
    labs(
      x = x_lab,
      y = y_lab
    ) +
    annotate(
      "text",
      x = -Inf, y = Inf,                          
      hjust = -0.05, vjust = 1.5,
      label = label_txt,
      size = 3.2
    ) +
    theme_bw() +
    theme(
      panel.grid = element_blank(),
      plot.margin = margin(5, 10, 5, 5)
    )
}

p_build <- make_scatter_lu_tp(
  data_plot,
  x_var = "Build_up_500",
  x_lab = "Urban (%)"
)

p_crops <- make_scatter_lu_tp(
  data_plot,
  x_var = "Crops_500",
  x_lab = "Agriculture (%)"
)

p_forest <- make_scatter_lu_tp(
  data_plot,
  x_var = "Forest_500",
  x_lab = "Forest (%)"
)

p_all <- (p_build | p_crops | p_forest)


ggsave(
  filename = "output_figure/Fig.S3.png",
  plot     = p_all,
  width    = 21,
  height   = 7,
  units    = "cm",
  dpi      = 300
)


```
